seperated definitions into respective classes
