+++ src/classifiers/anthropicClassifier.ts
@@ -1,22 +1,36 @@
 import {
-    ANTHROPIC_MODEL_ID_CLAUDE_3_5_SONNET,
-  ClassifierResult,
+  ANTHROPIC_MODEL_ID_CLAUDE_3_5_SONNET,
   ConversationMessage,
   ParticipantRole,
 } from "../types";
 import { isToolInput } from "../utils/helpers";
 import { Logger } from "../utils/logger";
-import { Classifier } from "./classifier";
+import { Classifier, ClassifierResult } from "./classifier";
 import { Anthropic } from "@anthropic-ai/sdk";
 
 export interface AnthropicClassifierOptions {
+  // Optional: The ID of the Anthropic model to use for classification
+  // If not provided, a default model may be used
   modelId?: string;
+
+  // Optional: Configuration for the inference process
   inferenceConfig?: {
+    // Maximum number of tokens to generate in the response
     maxTokens?: number;
+
+    // Controls randomness in output generation
+    // Higher values (e.g., 0.8) make output more random, lower values (e.g., 0.2) make it more deterministic
     temperature?: number;
+
+    // Controls diversity of output via nucleus sampling
+    // 1.0 considers all tokens, lower values (e.g., 0.9) consider only the most probable tokens
     topP?: number;
+
+    // Array of sequences that will stop the model from generating further tokens when encountered
     stopSequences?: string[];
   };
+
+  // The API key for authenticating with Anthropic's services
   apiKey: string;
 }
